Move Icon from primitives to ndla/icons
