+++ packages/ndla-icons/src/index.ts
@@ -8,3 +8,4 @@
 
 export { default as Spinner } from "./Spinner";
 export { default as Icon } from "./Icon";
+export type { Props as IconProps } from "./Icon";
